✨ feat: Import SiOpenai icon and replace 'Tag' component in chat feature
Import and use the SiOpenai icon from the '@icons-pack/react-simple-icons'
package. Replace instances of the 'Tag' component from the 'antd' package
with the 'Tag' component from the '@lobehub/ui' package in different
components within the chat feature of the application.

This commit introduces new features to enhance the chat feature by adding a new icon and updating the 'Tag' component.